[catalog-m3-Рефакторинг #88] - Implemented CollectionDetailsService.swift
